+++ docs/configuration.md
@@ -42,6 +42,9 @@ See the [cluster.Definition](../cluster/definition.go) and [cluster.Operator](..
 details on how this `cluster-definition.json` object is (de)serialised and how the [SSZ](https://github.com/ethereum/consensus-specs/blob/dev/ssz/simple-serialize.md)
 config hash and definition hash are calculated.
 
+The operator `config_signature` and `enr_signature` are [EIP712](https://eips.ethereum.org/EIPS/eip-712) signatures of **typed structured data** as opposed to just raw bytes. EIP712 enables users to see the object that they are signing in their wallet (ex: Metamask).
+See [eip712sigs.go](../cluster/eip712sigs.go) for details on the EIP712 structure used to create these signatures.
+
 The above `cluster-definition.json` is provided as input to the DKG which generates keys and the `cluster-lock.json` file.
 
 The `cluster-lock.json` has the following schema:
@@ -141,61 +144,3 @@ Flags:
 
 ````
 <!-- Code above generated by cmd/cmd_internal_test.go#TestConfigReference. DO NOT EDIT -->
-
-## EIP712 Signature Spec
-
-EIP712 is a procedure for hashing and signing of **typed structured data** as opposed to just bytestrings. EIP712 enables the user to see the object that they are signing in their wallet (ex: Metamask). For more details: [EIP712](https://eips.ethereum.org/EIPS/eip-712).
-We have 2 EIP712 signatures in cluster definition: config_signature and enr_signature.
-
-
-### EIP712 typed structure for config_signature:
-
-````go
-signerData := apitypes.TypedData {
-  Types: apitypes.Types {
-    "EIP712Domain": []apitypes.Type {
-      { Name: "name", Type: "string" }, // Name of protocol
-      { Name: "version", Type: "string" }, // Current version EIP712 Domain
-      { Name: "chainId", Type: "uint256" }, // Chain id of active chain from metamask.
-    },
-    "ConfigHash": []apitypes.Type {
-      { Name: "config_hash", Type: "string" }, // config_hash of cluster-definition.json
-    },
-  },
-  PrimaryType: "ConfigHash",
-  Domain: apitypes.TypedDataDomain {
-    Name:              "Obol",
-    Version:           "1",
-    ChainId:           ethmath.NewHexOrDecimal256(int64(chainID)),
-  },
-  Message: apitypes.TypedDataMessage {
-    "config_hash": configHash,
-  },
-}
-````
-
-### EIP712 typed structure for enr_signature:
-
-````go
-signerData := apitypes.TypedData {
-  Types: apitypes.Types {
-     "EIP712Domain": []apitypes.Type{
-        { Name: "name", Type: "string" },
-        { Name: "version", Type: "string" },
-        { Name: "chainId", Type: "uint256" },
-     },
-     "ENR": []apitypes.Type {
-        { Name: "enr", Type: "string" }, // ENR string of user.
-     },
-  },
-  PrimaryType: "ENR",
-  Domain: apitypes.TypedDataDomain {
-     Name:              "Obol",
-     Version:           "1",
-     ChainId:           ethmath.NewHexOrDecimal256(int64(chainID)),
-  },
-  Message: apitypes.TypedDataMessage {
-     "enr":     enr,
-  },
-}
-````
